Made MyException.h
